Scope quantity and custom-unit tests to mirrored modules
