+++ README.md
@@ -360,7 +360,7 @@ Integer cents are represented exactly throughout `[-Number.MAX_SAFE_INTEGER, Num
 
 The convenience `dollars` constructor intentionally accepts only whole dollars and checks the scaled result too; use `cents(425)` for $4.25. It is not a general decimal parser: multiplying arbitrary decimal numbers by 100 can introduce binary rounding errors. Parse decimal inputs exactly in your money library or with `Rational`, and choose an explicit policy for sub-cent amounts. `inDollars` is a numeric display crossing that may round, not a lossless storage or accounting boundary. Currency codes, symbols, locale formatting, decimal parsing, and rounding or allocation policies remain the application's responsibility.
 
-**Safe integer inputs do not make float arithmetic safe accounting.** `Quantity` does not preserve an integer or safe-range invariant: sums and products can overflow the safe range, and rates can produce fractional cents. For example, pricing one meter at 200 cents per three meters produces `200 / 3` cents, which `inCents` rejects. Even a final safe-integer result cannot prove intermediate computations were exact. Use [QuantityExact](#exact-quantities) or a money library as the system of record for accounting; exact rational arithmetic avoids float rounding and width limits but still needs an application policy for rounding to payable minor units. The executable recipes and rename regressions are in `test/CustomUnitRecipes.test.ts`; `test/CustomUnitsExact.test.ts` shows an exact money-library boundary.
+**Safe integer inputs do not make float arithmetic safe accounting.** `Quantity` does not preserve an integer or safe-range invariant: sums and products can overflow the safe range, and rates can produce fractional cents. For example, pricing one meter at 200 cents per three meters produces `200 / 3` cents, which `inCents` rejects. Even a final safe-integer result cannot prove intermediate computations were exact. Use [QuantityExact](#exact-quantities) or a money library as the system of record for accounting; exact rational arithmetic avoids float rounding and width limits but still needs an application policy for rounding to payable minor units. Custom-unit identity is tested in `test/Unit.test.ts`; quantity arithmetic, decimal crossings, and persistence are tested in `test/Quantity.test.ts`. `test/CustomUnitsExact.test.ts` shows an exact money-library boundary.
 
 ## Dimensionless quantities
