+++ test/Quantity.test.ts
@@ -1,4 +1,4 @@
-import { describe, expectTypeOf, it } from "@effect/vitest";
+import { describe, expect, expectTypeOf, it } from "@effect/vitest";
 import {
   assertEquals,
   assertFalse,
@@ -7,11 +7,14 @@ import {
   throws,
 } from "@effect/vitest/utils";
 import * as Array from "effect/Array";
+import * as BigDecimal from "effect/BigDecimal";
 import * as Result from "effect/Result";
 import * as Equal from "effect/Equal";
 import * as FastCheck from "effect/testing/FastCheck";
 import * as Function from "effect/Function";
+import * as Option from "effect/Option";
 import * as Schema from "effect/Schema";
+import * as SchemaTransformation from "effect/SchemaTransformation";
 
 import { isCloseTo, double } from "./testUtils.ts";
 import * as Dimensionless from "../src/Dimensionless.ts";
@@ -90,7 +93,169 @@ describe("point-free reducers", () => {
   });
 });
 
+describe("fromBigDecimal / toBigDecimal", () => {
+  it("supports direct and curried construction with precise units", () => {
+    const decimal = BigDecimal.fromStringUnsafe("1.25");
+    const direct = Quantity.fromBigDecimal("Meters", decimal);
+    const curried = decimal.pipe(Quantity.fromBigDecimal("Meters"));
+    expectTypeOf(direct).toEqualTypeOf<Quantity.Quantity<"Meters">>();
+    expectTypeOf(curried).toEqualTypeOf<Quantity.Quantity<"Meters">>();
+    expect(Quantity.equals(direct, curried)).toBe(true);
+    expect(direct.value).toBe(1.25);
+    const rate = Unit.rate(Unit.custom("USD"), Unit.custom("Count"));
+    const quantity = Quantity.fromBigDecimal(rate, decimal);
+    expectTypeOf(quantity).toEqualTypeOf<Quantity.Quantity<typeof rate>>();
+    expect(Unit.equals(quantity.unit, rate)).toBe(true);
+  });
+
+  it("round-trips up to 15 significant decimal digits in a normal finite range", () => {
+    FastCheck.assert(
+      FastCheck.property(
+        FastCheck.bigInt({
+          min: -999_999_999_999_999n,
+          max: 999_999_999_999_999n,
+        }),
+        FastCheck.integer({ min: -293, max: 307 }),
+        (coefficient, scale) => {
+          const decimal = BigDecimal.make(coefficient, scale);
+          const quantity = Quantity.fromBigDecimal("Meters", decimal);
+          const recovered = Option.getOrThrow(Quantity.toBigDecimal(quantity));
+          expect(BigDecimal.equals(recovered, decimal)).toBe(true);
+        },
+      ),
+      { numRuns: 5000 },
+    );
+  });
+
+  it("handles overflow, underflow, and precision beyond the round-trip guarantee", () => {
+    for (const [input, value] of [
+      ["1e309", Infinity],
+      ["-1e309", -Infinity],
+      ["1e-400", 0],
+      ["-1e-400", 0],
+      ["9007199254740993", 9007199254740992],
+      ["9007199254740995", 9007199254740996],
+    ] as const) {
+      expect(
+        Quantity.fromBigDecimal("Meters", BigDecimal.fromStringUnsafe(input))
+          .value,
+      ).toBe(value);
+    }
+    const subnormal = BigDecimal.fromStringUnsafe("1.23456789012345e-310");
+    const recovered = Option.getOrThrow(
+      Quantity.toBigDecimal(Quantity.fromBigDecimal("Meters", subnormal)),
+    );
+    expect(BigDecimal.equals(recovered, subnormal)).toBe(false);
+    for (const value of [NaN, Infinity, -Infinity]) {
+      expect(
+        Option.isNone(Quantity.toBigDecimal(Quantity.make("Meters", value))),
+      ).toBe(true);
+    }
+  });
+
+  it("round-trips normal-range boundary decimals and finite double inputs", () => {
+    for (const input of [
+      "2.22507385850721e-308",
+      "-2.22507385850721e-308",
+      "1.79769313486231e308",
+      "-1.79769313486231e308",
+      "0.000",
+      "1.2300",
+    ]) {
+      const decimal = BigDecimal.fromStringUnsafe(input);
+      const recovered = Option.getOrThrow(
+        Quantity.toBigDecimal(Quantity.fromBigDecimal("Meters", decimal)),
+      );
+      expect(BigDecimal.equals(recovered, decimal)).toBe(true);
+    }
+    FastCheck.assert(
+      FastCheck.property(
+        FastCheck.double({ noNaN: true, noDefaultInfinity: true }),
+        (value) => {
+          const quantity = Quantity.make("Meters", value);
+          const decimal = Option.getOrThrow(Quantity.toBigDecimal(quantity));
+          expect(Quantity.fromBigDecimal("Meters", decimal).value).toBe(
+            quantity.value,
+          );
+        },
+      ),
+      { numRuns: 1000 },
+    );
+  });
+
+  it("uses shortest decimal printing rather than the exact binary fraction", () => {
+    const decimal = Option.getOrThrow(
+      Quantity.toBigDecimal(Quantity.make("Meters", 0.1)),
+    );
+    expect(BigDecimal.equals(decimal, BigDecimal.make(1n, 1))).toBe(true);
+    const computed = Option.getOrThrow(
+      Quantity.toBigDecimal(Quantity.make("Meters", 0.1 + 0.2)),
+    );
+    expect(BigDecimal.equals(computed, BigDecimal.make(3n, 1))).toBe(false);
+  });
+
+  it("shifts a decimal scale before rounding for cents-backed custom units", () => {
+    const Usd = Unit.custom("USD");
+    const dollars = BigDecimal.fromStringUnsafe("0.29");
+    const cents = Quantity.fromBigDecimal(
+      Usd,
+      BigDecimal.make(dollars.value, dollars.scale - 2),
+    );
+    expect(cents.value).toBe(29);
+    expect(0.29 * 100).not.toBe(29);
+    const recoveredCents = Option.getOrThrow(Quantity.toBigDecimal(cents));
+    const recoveredDollars = BigDecimal.make(
+      recoveredCents.value,
+      recoveredCents.scale + 2,
+    );
+    expect(BigDecimal.equals(recoveredDollars, dollars)).toBe(true);
+  });
+
+  it("converts custom-unit values to decimals without display scaling", () => {
+    const Usd = Unit.custom("USD");
+    const output = Quantity.toBigDecimal(Quantity.make(Usd, 425));
+
+    expect(
+      BigDecimal.equals(Option.getOrThrow(output), BigDecimal.make(425n, 0)),
+    ).toBe(true);
+  });
+});
+
+describe("make", () => {
+  it("stores numeric values without imposing an integer or safe-range invariant", () => {
+    const unit = Unit.custom("USD");
+    for (const value of [
+      0,
+      -425,
+      Number.MAX_SAFE_INTEGER,
+      -Number.MAX_SAFE_INTEGER,
+      0.5,
+      Number.MAX_SAFE_INTEGER + 1,
+      -Number.MAX_SAFE_INTEGER - 1,
+      NaN,
+      Infinity,
+      -Infinity,
+    ]) {
+      const quantity = Quantity.make(unit, value);
+      assertEquals(quantity.value, value);
+      assertTrue(Unit.equals(quantity.unit, unit));
+    }
+  });
+});
+
 describe("multiply", () => {
+  it("can exceed the safe-integer range from a safe-integer input", () => {
+    const unit = Unit.custom("USD");
+    const result = Quantity.multiply(
+      Quantity.make(unit, Number.MAX_SAFE_INTEGER),
+      2,
+    );
+
+    assertEquals(result.value, Number.MAX_SAFE_INTEGER * 2);
+    assertFalse(Number.isSafeInteger(result.value));
+    assertTrue(Unit.equals(result.unit, unit));
+  });
+
   const baseQuantities = [
     { label: "Length", constructor: Length.meters },
     { label: "Mass", constructor: Mass.kilograms },
@@ -193,6 +358,32 @@ describe("squared/cubed", () => {
 
 describe("rates", () => {
   const seconds = (n: number) => Quantity.make("Seconds", n);
+
+  it("per and at preserve custom-unit base values without display scaling", () => {
+    const unit = Unit.custom("USD");
+    const price = Quantity.per(
+      Quantity.make(unit, 300),
+      Quantity.make("Meters", 2),
+    );
+    const cost = Quantity.at(price, Quantity.make("Meters", 10));
+
+    assertEquals(price.value, 150);
+    assertTrue(Unit.equals(price.unit, Unit.rate(unit, "Meters")));
+    assertEquals(cost.value, 1500);
+    assertTrue(Unit.equals(cost.unit, unit));
+  });
+
+  it("can produce fractional custom-unit values from integer inputs", () => {
+    const unit = Unit.custom("USD");
+    const result = Quantity.at(
+      Quantity.per(Quantity.make(unit, 200), Quantity.make("Meters", 3)),
+      Quantity.make("Meters", 1),
+    );
+
+    assertEquals(result.value, 200 / 3);
+    assertFalse(Number.isSafeInteger(result.value));
+    assertTrue(Unit.equals(result.unit, unit));
+  });
 
   it("per divides values and forms a Rate unit", () => {
     FastCheck.assert(
@@ -472,6 +663,418 @@ describe("dimensionless", () => {
 });
 
 describe("schema", () => {
+  const Count = Unit.custom("Count");
+  const count = (value: number) => Quantity.make(Count, value);
+
+  describe("QuantityFromValue", () => {
+    it("round-trips base, custom, and derived units as bare numbers", () => {
+      for (const unit of [
+        "Meters",
+        Count,
+        Unit.rate(Count, "Seconds"),
+      ] as const) {
+        const codec = Quantity.QuantityFromValue(unit);
+        for (const value of [0, -1, 1.5, Number.MAX_VALUE, Number.MIN_VALUE]) {
+          const quantity = Schema.decodeUnknownSync(codec)(value);
+          expect(Quantity.equals(quantity, Quantity.make(unit, value))).toBe(
+            true,
+          );
+          expect(Schema.encodeSync(codec)(quantity)).toBe(value);
+        }
+      }
+    });
+
+    it("rejects non-finite values, invalid wires, and the wrong unit", () => {
+      const codec = Quantity.QuantityFromValue(Count);
+      for (const value of [NaN, Infinity, -Infinity, "1", null, { value: 1 }]) {
+        expect(Result.isFailure(Schema.decodeUnknownResult(codec)(value))).toBe(
+          true,
+        );
+      }
+      for (const value of [NaN, Infinity, -Infinity]) {
+        expect(Result.isFailure(Schema.encodeResult(codec)(count(value)))).toBe(
+          true,
+        );
+      }
+      expect(
+        Result.isFailure(Schema.encodeUnknownResult(codec)(Length.meters(1))),
+      ).toBe(true);
+    });
+
+    it("preserves the precise types and explicit bare-value JSON codec", () => {
+      const codec = Quantity.QuantityFromValue(Count);
+      expectTypeOf<typeof codec.Type>().toEqualTypeOf<
+        Quantity.Quantity<typeof Count>
+      >();
+      expectTypeOf<typeof codec.Encoded>().toEqualTypeOf<number>();
+      const json = Schema.toCodecJson(Schema.Struct({ count: codec }));
+      expect(Schema.encodeSync(json)({ count: count(2) })).toEqual({
+        count: 2,
+      });
+      expect(Schema.decodeUnknownSync(json)({ count: 2 }).count.value).toBe(2);
+      const canonical = Schema.toCodecJson(Quantity.Quantity(Count));
+      expect(Schema.encodeSync(canonical)(count(2))).toEqual({
+        unit: "[Count]",
+        value: 2,
+      });
+    });
+
+    it("normalizes negative zero", () => {
+      const codec = Quantity.QuantityFromValue(Count);
+      expect(Object.is(Schema.decodeUnknownSync(codec)(-0).value, 0)).toBe(
+        true,
+      );
+    });
+
+    it("composes numeric wire checks before the quantity conversion", () => {
+      const codec = Schema.Int.check(Schema.isGreaterThan(0)).pipe(
+        Schema.decodeTo(Quantity.QuantityFromValue(Count)),
+      );
+      expect(Schema.decodeUnknownSync(codec)(2).value).toBe(2);
+      expect(Schema.encodeSync(codec)(count(2))).toBe(2);
+      for (const value of [0, -1, 1.5]) {
+        expect(Result.isFailure(Schema.decodeUnknownResult(codec)(value))).toBe(
+          true,
+        );
+        expect(Result.isFailure(Schema.encodeResult(codec)(count(value)))).toBe(
+          true,
+        );
+      }
+    });
+  });
+
+  describe("quantity refinements", () => {
+    it("preserves existing arbitrary annotations while applying checks", () => {
+      const schema = Quantity.positive(
+        Length.Length.annotate(
+          Quantity.arbitraryOnGrid("Meters", { step: 0.1, min: -1, max: 1 }),
+        ),
+      );
+      FastCheck.assert(
+        FastCheck.property(Schema.toArbitrary(schema), (quantity) => {
+          expect(quantity.value).toBeGreaterThan(0);
+          expect(quantity.value).toBeLessThanOrEqual(1);
+          expect(quantity.value).toBe(Number(quantity.value.toFixed(1)));
+          expect(Unit.equals(quantity.unit, "Meters")).toBe(true);
+        }),
+      );
+    });
+
+    it("checks positivity and non-negativity without changing wire types", () => {
+      const positive = Quantity.positive(Quantity.QuantityFromValue(Count));
+      const nonNegative = Quantity.nonNegative(
+        Quantity.QuantityFromStruct(Count),
+      );
+      expectTypeOf<typeof positive.Encoded>().toEqualTypeOf<number>();
+      expectTypeOf<typeof nonNegative.Encoded>().toEqualTypeOf<{
+        readonly unit: string;
+        readonly value: number;
+      }>();
+      expect(Schema.decodeUnknownSync(positive)(1).value).toBe(1);
+      expect(Schema.encodeSync(positive)(count(1))).toBe(1);
+      for (const value of [0, -0, -1]) {
+        expect(
+          Result.isFailure(Schema.decodeUnknownResult(positive)(value)),
+        ).toBe(true);
+        expect(
+          Result.isFailure(Schema.encodeResult(positive)(count(value))),
+        ).toBe(true);
+      }
+      for (const value of [0, -0, 1]) {
+        expect(
+          Schema.decodeUnknownSync(nonNegative)({ unit: "[Count]", value })
+            .value,
+        ).toBe(value === 0 ? 0 : value);
+      }
+      expect(
+        Result.isFailure(Schema.encodeResult(nonNegative)(count(-1))),
+      ).toBe(true);
+    });
+
+    it("uses quantity comparisons for built-in and derived units", () => {
+      const minimumLength = Length.LengthFromStruct.check(
+        Quantity.greaterThanOrEqualTo(Length.centimeters(100)),
+      );
+      expect(Schema.encodeSync(minimumLength)(Length.meters(1)).value).toBe(1);
+      expect(
+        Result.isFailure(
+          Schema.encodeResult(minimumLength)(Length.centimeters(99)),
+        ),
+      ).toBe(true);
+      const rate = Unit.rate(Count, "Seconds");
+      const codec = Quantity.QuantityFromValue(rate).check(
+        Quantity.greaterThanOrEqualTo(Quantity.make(rate, 2)),
+      );
+      expect(Schema.decodeUnknownSync(codec)(2).value).toBe(2);
+      expect(Result.isFailure(Schema.decodeUnknownResult(codec)(1))).toBe(true);
+    });
+
+    it("rejects mismatched bound units statically and dynamically", () => {
+      const codec = Quantity.QuantityFromValue(Count);
+      if (globalThis.Boolean(false)) {
+        // @ts-expect-error A length bound cannot refine counts.
+        codec.check(Quantity.greaterThanOrEqualTo(Length.meters(1)));
+        // @ts-expect-error A number schema does not decode quantities.
+        Quantity.positive(Schema.Number);
+      }
+      const broadBound: Quantity.Quantity<Unit.Unit> = Length.meters(1);
+      const checked = codec.check(Quantity.greaterThanOrEqualTo(broadBound));
+      expect(Result.isFailure(Schema.decodeUnknownResult(checked)(2))).toBe(
+        true,
+      );
+    });
+
+    it("reports unit-aware defaults and supports custom messages", () => {
+      const positive = Quantity.positive(Quantity.Quantity(Count));
+      const nonNegative = Quantity.nonNegative(Quantity.Quantity(Count));
+      const bounded = Quantity.Quantity(Count).check(
+        Quantity.greaterThanOrEqualTo(count(1)),
+      );
+      expect(() => Schema.decodeUnknownSync(positive)(count(0))).toThrow(
+        "positive quantity in [Count]",
+      );
+      expect(() => Schema.decodeUnknownSync(nonNegative)(count(-1))).toThrow(
+        "non-negative quantity in [Count]",
+      );
+      expect(() => Schema.decodeUnknownSync(bounded)(count(0))).toThrow(
+        "greater than or equal to 1 [Count]",
+      );
+      for (const codec of [
+        Quantity.positive(Quantity.Quantity(Count), {
+          message: "A count is required",
+        }),
+        Quantity.nonNegative(Quantity.Quantity(Count), {
+          message: "A count is required",
+        }),
+        Quantity.Quantity(Count).check(
+          Quantity.greaterThanOrEqualTo(count(1), {
+            message: "A count is required",
+          }),
+        ),
+      ]) {
+        expect(() => Schema.decodeUnknownSync(codec)(count(-1))).toThrow(
+          "A count is required",
+        );
+      }
+    });
+
+    it("follows IEEE comparisons while JSON encoding remains finite-only", () => {
+      const positive = Quantity.positive(Quantity.Quantity(Count));
+      const nonNegative = Quantity.nonNegative(Quantity.Quantity(Count));
+      const bounded = Quantity.Quantity(Count).check(
+        Quantity.greaterThanOrEqualTo(count(1)),
+      );
+      for (const schema of [positive, nonNegative, bounded]) {
+        expect(Schema.is(schema)(count(NaN))).toBe(false);
+        expect(Schema.is(schema)(count(-Infinity))).toBe(false);
+        expect(Schema.is(schema)(count(Infinity))).toBe(true);
+        const json = Schema.toCodecJson(Schema.Struct({ count: schema }));
+        expect(
+          Result.isFailure(
+            Schema.encodeResult(json)({ count: count(Infinity) }),
+          ),
+        ).toBe(true);
+        expect(
+          Result.isFailure(
+            Schema.decodeUnknownResult(json)({
+              count: { unit: "[Count]", value: -1 },
+            }),
+          ),
+        ).toBe(true);
+        expect(Schema.encodeSync(json)({ count: count(1) })).toEqual({
+          count: { unit: "[Count]", value: 1 },
+        });
+      }
+    });
+  });
+
+  describe("wrapping an existing schema", () => {
+    const input = Schema.NumberFromString.check(
+      Schema.isInt(),
+      Schema.isGreaterThan(0, { message: "Enter a positive whole count" }),
+    );
+    const codec = input.pipe(
+      Schema.decodeTo(
+        Quantity.Quantity(Count),
+        SchemaTransformation.transform({
+          decode: count,
+          encode: (quantity) => quantity.value,
+        }),
+      ),
+    );
+
+    it("preserves base validation and messages in both directions", () => {
+      expect(Schema.decodeUnknownSync(codec)("2").value).toBe(2);
+      expect(Schema.encodeSync(codec)(count(2))).toBe("2");
+      expect(() => Schema.decodeUnknownSync(codec)("0")).toThrow(
+        "Enter a positive whole count",
+      );
+      expect(() => Schema.encodeSync(codec)(count(0))).toThrow(
+        "Enter a positive whole count",
+      );
+      expect(Result.isFailure(Schema.decodeUnknownResult(codec)("1.5"))).toBe(
+        true,
+      );
+      expect(Result.isFailure(Schema.encodeResult(codec)(count(1.5)))).toBe(
+        true,
+      );
+    });
+
+    it("retains existing checks and schema types when refined", () => {
+      const refined = Quantity.positive(codec);
+      expectTypeOf<typeof refined.Encoded>().toEqualTypeOf<string>();
+      expectTypeOf<typeof refined.Type>().toEqualTypeOf<
+        Quantity.Quantity<typeof Count>
+      >();
+      expectTypeOf<typeof refined.DecodingServices>().toEqualTypeOf<
+        typeof codec.DecodingServices
+      >();
+      expectTypeOf<typeof refined.EncodingServices>().toEqualTypeOf<
+        typeof codec.EncodingServices
+      >();
+      expect(Result.isFailure(Schema.decodeUnknownResult(refined)("1.5"))).toBe(
+        true,
+      );
+      expect(Schema.encodeSync(refined)(count(2))).toBe("2");
+    });
+
+    it("composes refined quantity codecs with existing encoded representations", () => {
+      const PositiveMeters = Quantity.positive(
+        Quantity.QuantityFromValue("Meters"),
+      );
+      const NonNegativeLength = Quantity.nonNegative(Length.LengthFromStruct);
+      const AtLeastOneMeter = PositiveMeters.check(
+        Quantity.greaterThanOrEqualTo(Length.meters(1)),
+      );
+      expect(
+        Quantity.equals(
+          Schema.decodeSync(AtLeastOneMeter)(2),
+          Length.meters(2),
+        ),
+      ).toBe(true);
+      expect(Schema.encodeSync(AtLeastOneMeter)(Length.meters(2))).toBe(2);
+      expect(Schema.encodeSync(NonNegativeLength)(Length.meters(0))).toEqual({
+        unit: "Meters",
+        value: 0,
+      });
+
+      const ExistingMeters = Schema.NumberFromString.check(
+        Schema.isFinite(),
+        Schema.isGreaterThan(0, { expected: "a positive distance in meters" }),
+      );
+      const Distance = ExistingMeters.pipe(
+        Schema.decodeTo(
+          Quantity.Quantity("Meters"),
+          SchemaTransformation.transform({
+            decode: (value) => Quantity.make("Meters", value),
+            encode: (quantity) => quantity.value,
+          }),
+        ),
+      );
+      expect(
+        Quantity.equals(
+          Schema.decodeSync(Distance)("2"),
+          Quantity.make("Meters", 2),
+        ),
+      ).toBe(true);
+      expect(Schema.encodeSync(Distance)(Quantity.make("Meters", 2))).toBe("2");
+      expect(() => Schema.decodeSync(Distance)("0")).toThrow(
+        "a positive distance in meters",
+      );
+      expect(() =>
+        Schema.encodeSync(Distance)(Quantity.make("Meters", 0)),
+      ).toThrow("a positive distance in meters");
+      expect(
+        Result.isFailure(Schema.decodeUnknownResult(Distance)("Infinity")),
+      ).toBe(true);
+    });
+  });
+
+  it("round-trips custom-unit values without display scaling", () => {
+    const unit = Unit.custom("USD");
+    const codec = Quantity.QuantityFromStruct(unit);
+    const quantity = Quantity.make(unit, 1500);
+    const encoded = Schema.encodeSync(codec)(quantity);
+
+    deepStrictEqual(encoded, { unit: "[USD]", value: 1500 });
+    assertTrue(Quantity.equals(Schema.decodeSync(codec)(encoded), quantity));
+  });
+
+  it("accepts finite fractional and unsafe-integer values without a cents invariant", () => {
+    const unit = Unit.custom("USD");
+    const codec = Quantity.QuantityFromStruct(unit);
+
+    for (const value of [0.5, Number.MAX_SAFE_INTEGER + 1]) {
+      const encoded = { unit: "[USD]", value };
+      const decoded = Schema.decodeSync(codec)(encoded);
+
+      assertEquals(decoded.value, value);
+      assertTrue(Unit.equals(decoded.unit, unit));
+      deepStrictEqual(Schema.encodeSync(codec)(decoded), encoded);
+    }
+  });
+
+  it("accepts a retained custom id and rejects a changed id", () => {
+    const original = Unit.custom("Units");
+    const renamed = Unit.custom("Units");
+    const stored = Schema.encodeSync(Quantity.QuantityFromStruct(original))(
+      Quantity.make(original, 3),
+    );
+    const codec = Quantity.QuantityFromStruct(renamed);
+    const decoded = Schema.decodeSync(codec)(stored);
+
+    assertTrue(Quantity.equals(decoded, Quantity.make(renamed, 3)));
+    deepStrictEqual(Schema.encodeSync(codec)(decoded), {
+      unit: "[Units]",
+      value: 3,
+    });
+    assertTrue(
+      Result.isFailure(
+        Schema.decodeUnknownResult(
+          Quantity.QuantityFromStruct(Unit.custom("Count")),
+        )(stored),
+      ),
+    );
+  });
+
+  it("enforces custom ids inside derived rate and product wires", () => {
+    const original = Unit.custom("Units");
+    const renamed = Unit.custom("Units");
+    const changed = Unit.custom("Count");
+    const usd = Unit.custom("USD");
+
+    for (const [before, after, different, wire] of [
+      [
+        Unit.rate(usd, original),
+        Unit.rate(usd, renamed),
+        Unit.rate(usd, changed),
+        "([USD]/[Units])",
+      ],
+      [
+        Unit.product(original, "Meters"),
+        Unit.product(renamed, "Meters"),
+        Unit.product(changed, "Meters"),
+        "([Units]*Meters)",
+      ],
+    ] as const) {
+      const stored = Schema.encodeSync(Quantity.QuantityFromStruct(before))(
+        Quantity.make(before, 2),
+      );
+      deepStrictEqual(stored, { unit: wire, value: 2 });
+      const decoded = Schema.decodeSync(Quantity.QuantityFromStruct(after))(
+        stored,
+      );
+      assertTrue(Quantity.equals(decoded, Quantity.make(after, 2)));
+      assertTrue(
+        Result.isFailure(
+          Schema.decodeUnknownResult(Quantity.QuantityFromStruct(different))(
+            stored,
+          ),
+        ),
+      );
+    }
+  });
+
   it("derives grid-constrained quantities from annotations", () => {
     const MeasuredLength = Length.Length.annotate(
       Quantity.arbitraryOnGrid(Length.Meters, {
